[CHERI] Fix dynamic stack alloc to allow alignment > size
Introduces an additional mask operation when dynamically allocating
stack, on top of the existing mask generated by the cram - which
appears to be solely about generating a mask which forces the alignment
which would be necessary for the compressed capability format, based on
the requested allocation size. However it doesn't take into account the
additional requested alignment, which may be greater than that required
by cram.