[InstrProfiling] Support profile/coverage generation for CHERI targets
On CHERI, enabling instruction profiling (via GCov etc.) involves
supporting the runtime generation of profiling data via compiler-
inserted intrinsics, and the exporting of that data in a format that can
be processed by non-CHERI targets. These changes include the necessary
modifications to those intrinsics to make their signature compatible
with CHERI capabilities, modifications to the data structures involved
in runtime accumulation of profiling data to use capabilities on CHERI,
and also modifications to the logic of the accumulation and exporting
processes. It was necessary to use two data formats for the profiling
structure - one to be used internally at runtime, which had the space
and layout necessary to use capabilities to represent pointers, and
another to be stored as the exported profiling data, which uses the raw
address part of the pointers such that it can then be processed
correctly by a non-CHERI target. This implied a step whereby the
internal format needs to be converted into the exported format.